djchat/requirements.txt
anyio==4.2.0
asgiref==3.7.2
asttokens==2.4.1
attrs==23.1.0
backcall==0.2.0
backports.zoneinfo==0.2.1
black==23.10.0
channels==4.0.0
click==8.1.7
decorator==5.1.1
Django==4.2.6
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.26.5
exceptiongroup==1.2.0
executing==2.0.1
flake8==6.1.0
h11==0.14.0
httptools==0.6.1
idna==3.6
importlib-resources==6.1.0
inflection==0.5.1
ipython==8.12.3
jedi==0.19.1
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
matplotlib-inline==0.1.6
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.2
parso==0.8.3
pathspec==0.11.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
pkgutil-resolve-name==1.3.10
platformdirs==3.11.0
prompt-toolkit==3.0.39
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.11.1
pyflakes==3.1.0
Pygments==2.16.1
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
rpds-py==0.10.6
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
stack-data==0.6.3
tomli==2.0.1
traitlets==5.13.0
typing-extensions==4.8.0
uritemplate==4.1.1
uvicorn==0.25.0
uvloop==0.19.0
watchfiles==0.21.0
wcwidth==0.2.9
websockets==12.0
zipp==3.17.0
